+++ runescape/src/main/java/Class151.java
@@ -32,7 +32,7 @@ public final class Class151 {
 
     @OriginalMember(owner = "client!gt", name = "a", descriptor = "(BI)Lclient!bq;")
     public DoublyLinkedNode_Sub2_Sub3 method3234(@OriginalArg(1) int arg0) {
-        @Pc(11) DoublyLinkedNode_Sub2_Sub3 local11 = (DoublyLinkedNode_Sub2_Sub3) this.aReferenceCache_79.get((long) arg0);
+        @Pc(11) DoublyLinkedNode_Sub2_Sub3 local11 = (DoublyLinkedNode_Sub2_Sub3) this.aReferenceCache_79.get(arg0);
         if (local11 != null) {
             return local11;
         }
@@ -49,7 +49,7 @@ public final class Class151 {
         if (arg0 >= 32768) {
             local11.method1186();
         }
-        this.aReferenceCache_79.put(local11, (long) arg0);
+        this.aReferenceCache_79.put(local11, arg0);
         return local11;
     }
 }
